requirements.txt
aiogram==3.4.1
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.3
python-dotenv==1.0.1
sqlmodel==0.0.16
aiosqlite>=0.19
alembic==1.13.1